requirements.txt
cutie
psutil
requests
rich
PyYAML
